ROR-825 Add support for holes in sosi polygons
Interpreting SosiRefIsland to be a hole within a polygon. If present,
create a polygon using SosiRefNumber coordinates as shell and
coordinates of any SosiRefIsland instances as holes.

Put the final polygon into a multipolygon. Not sure if this last part is
required, but it seems most examples of polygons with holes are put into
mulitpolygons.